Create opecallout.js
Any custom code below to use the response from callout and use it in the CRM.


The provided Zoho CRM client script code is used to open a callout widget within Zoho CRM. Here's a description of the code and an explanation of its functionality:

The code starts by declaring a variable named `callout_response` which will store the response returned by the `ZDK.Client.openCallout()` function. The `openCallout()` function takes three parameters: 

1. Widget Configuration: It specifies the configuration details of the callout widget, including its API name, type (widget), header text, height, and width.

2. Field Configuration: This parameter defines the field details related to the callout, such as the field's API name.

3. Data: This parameter allows passing any additional data that needs to be sent to the callout widget.

Once the `openCallout()` function is executed, the response is logged to the console using `console.log(callout_response)`.

This client script code can be triggered on various events within Zoho CRM. The specific events depend on where and how the script is implemented. Some possible events include:
- Form submission: The script can be triggered when a specific form is submitted by a user.
- Field change: The script can execute when a particular field value is modified.
- Field On Type: The script can execute when a typing on for the field value.
- Page load: The script can run when a particular page or module is loaded.

The code can be customized to be triggered on different events based on your specific requirements within Zoho CRM.